+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation_objc/include/camera_avfoundation/FLTCam.h
@@ -56,7 +56,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)setDeviceOrientation:(UIDeviceOrientation)orientation;
 - (void)captureToFileWithCompletion:(void (^)(NSString *_Nullable,
                                               FlutterError *_Nullable))completion;
-- (void)setImageFileFormat:(FCPPlatformImageFileFormat)fileFormat;
 /// Starts recording a video with an optional streaming messenger.
 /// If the messenger is non-nil then it will be called for each
 /// captured frame, allowing streaming concurrently with recording.